Added a LINQtoSQL capabilities for UserLogin as well as ExceptionLogging.
Added userGroups eniumerator to handle the user grouping.

Added externsion methods to handle converting a string to a enum as well as writing , displaying and saving exceptions to the server.

Added a SQLCommunicator object to handle all the communications between the application and the sqlserver.

Added a ErrorLoggerObject to Handle all the error logging functions.

Added an ApplicationUser Object to save the user details of the current user.