Attempting to connect to the Barcode Scanner as a COM port device.
So far detecting the corrent COM port for the device is proving difficult as it does not show up in the WIN32_SerialPort search.

I found That it does show up in the Win32_PnPEntity search. I now need to find out how to extract the Name or the Caption of the object that is returned.