Added a Serial Communicator Object to handle finding, connecting, and recieving data from the barcode scanner.
Added a method in the CommandClient View Model to use the SerialCommunicator Object and pass the scanned Barcodes to the CommandServer.

Added a BarcodeScanner Obect so we can track all the information if multiple barcode scanners are connected.

Modified the BarcodeObject Definition to include the newly Added Group Name.

Modified the BarcodeDecoder Object to correctly decode the newbarcode standard like "*None+S1@OK"

Modified the VCS Communicator to add the byte array in the way that is expected according to VCS team's requirements. Right now accoding to the DLL the byte array is 4 bytes long.

Issues:

At the moment the UI of the CommandClient does not update when a barcode is scanned using the COMPort communication feature. Need to fix this bug to improve application usability.