Removed the Task.Run that was being used to capture and send the scanner values through the COM port to the CommandServer.
Replaced it with an AsyncAwait methodology; now the UI is updating flawlessly.

Removed BarcodeType and added Status to the Barcode Object.

Updated the code in the BarcodeDecoder object to get the Destination and Command from the table in the SQL Server.

Updated the GetDesdtination() and GetCommand() methods in the SQLCOmmunicator class to check against the Station name as well as the status.