Modified the code in the CommandClient to support many barcodescanners.
Heavily Modified and reafactored the SerialCommunicatorObject to be able to connect to and communicate with many barcode scanners.

Added a BarcodeScannerConfig Object to hold the scanner configurations to add those scanners to the list of ports we are reading.

Added a method in the SQL Communicator class to provide us with hard coded stgrings so the functions can be tested.

At the moment I have an Isse, one scanner works fine, but when I introduce the other I am getting many IO Errors. I will look into this and try to solve the issue.