Added a method to finalize the COM Ports when leaving the CommandClient.
Added a string Extension to remove extra characters gotten when reading from the Barcode Scanner.

Modified the barcode decoder to use the EleiminateExtraChars Extension method to remove any junk chars from the Command.

Modified the definition of the Barcode Scanner object to hold a Serial Port object that it is connected to.

Modified the SerialCommunicator Object by removing the common SerialPortObj and replacing it with a new object for every instace of Scanner that is connected.