Converted all the static objects to state objects in the Command Client.
Created an object to handle the creation of tasks with the capability of creating tasks in Queue.

Added barcodeDevoding, sending command to VCS and task cration capabilitiers to the Server.

Added some methods to the sql communicator object to support the creation of the task.

updated the barcode decoder object to a state object.

Added as tring extension to check if a barcode has all its data.