Modified the AGVControlCenter Server to support sending commands to the VCS as well as creating tasks.
Refactored the VCSCommunicator by making one common method to send messages, one to code messages and one to build the specific message we want.

Modified the SQL Communicator to always get the station data directly from the SQL server and not the cache.

Added a method inside the agvTaskCreator that will dequeue tasks.

Added a method in the ServerViewModel to start the dequeuing process.